+++ build.mjs
@@ -24,7 +24,7 @@ await Promise.all(
       platform: 'node',
       target: 'node18',
       outfile,
-      external: ['better-sqlite3'],
+      external: ['node:sqlite'],
     }).then(() => addShebang(outfile))
   )
 )